// packages/notation-model/src/mnx/types.ts
export type MnxDocument = GlobalAttrs & {
  global: Global;
  layouts?: SystemLayout[];
  mnx: Mnx;
  parts: Part[];
  scores?: Score[];
};
export type String = string;
export type Id = string;
export type Global = GlobalAttrs & {
  lyrics?: LyricsGlobal;
  measures: MeasureGlobal[];
  sounds?: SoundsGlobal;
};
export type LyricsGlobal = GlobalAttrs & {
  lineMetadata?: LyricLinesMetadata;
  lineOrder?: LyricLineId[];
};
/**
 * This interface was referenced by `LyricLinesMetadata`'s JSON-Schema definition
 * via the `patternProperty` "^.*$".
 */
export type LyricLineMetadata = GlobalAttrs & {
  label?: LyricLineLabel;
  lang?: LanguageCode;
};
export type LyricLineLabel = string;
export type LanguageCode = string;
export type LyricLineId = string;
export type MeasureGlobal = GlobalAttrs & {
  barline?: Barline;
  ending?: Ending;
  fermata?: Fermata;
  fine?: Fine;
  jump?: Jump;
  key?: Key;
  number?: MeasureNumber;
  repeatEnd?: RepeatEnd;
  repeatStart?: RepeatStart;
  segno?: Segno;
  tempos?: Tempo[];
  time?: Time;
};
export type Barline = GlobalAttrs & {
  type: BarlineType;
};
export type BarlineType =
  | 'regular'
  | 'dotted'
  | 'dashed'
  | 'heavy'
  | 'double'
  | 'final'
  | 'heavyLight'
  | 'heavyHeavy'
  | 'tick'
  | 'short'
  | 'noBarline';
export type Ending = GlobalAttrs & {
  color?: Color;
  duration: EndingDuration;
  numbers?: EndingNumber[];
  open?: EndingOpen;
};
export type Color = string;
export type EndingDuration = number;
export type EndingNumber = number;
export type EndingOpen = boolean;
export type Fermata = GlobalAttrs & {
  duration?: FermataDuration;
  placement?: Placement;
  pointing?: UpDownAuto;
  symbol?: FermataSymbol;
};
export type FermataDuration = 'auto' | 'none' | 'veryLong' | 'long' | 'normal' | 'short' | 'veryShort';
export type Placement = 'above' | 'below' | 'auto';
export type UpDownAuto = 'up' | 'down' | 'auto';
export type FermataSymbol =
  'normal' | 'angled' | 'square' | 'doubleAngled' | 'doubleSquare' | 'doubleDot' | 'halfCurve' | 'curlew';
export type Fine = GlobalAttrs & {
  color?: Color;
  location: RhythmicPosition;
};
export type RhythmicPosition = GlobalAttrs & {
  fraction: Fraction;
  graceIndex?: IntegerUnsigned;
};
/**
 * @minItems 2
 * @maxItems 2
 */
export type Fraction = [IntegerUnsigned, IntegerUnsigned];
export type IntegerUnsigned = number;
export type Jump = GlobalAttrs & {
  location: RhythmicPosition;
  type: JumpType;
};
export type JumpType = 'dsalfine' | 'segno';
export type Key = GlobalAttrs & {
  color?: Color;
  fifths: Fifths;
};
export type Fifths = number;
export type MeasureNumber = number;
export type RepeatEnd = GlobalAttrs & {
  times?: RepeatTimes;
};
export type RepeatTimes = number;
export type RepeatStart = GlobalAttrs & {};
export type Segno = GlobalAttrs & {
  color?: Color;
  glyph?: SmuflGlyph;
  location: RhythmicPosition;
};
export type SmuflGlyph = string;
export type Tempo = GlobalAttrs & {
  bpm: Bpm;
  location?: RhythmicPosition;
  value: NoteValue;
};
export type Bpm = number;
export type NoteValue = GlobalAttrs & {
  base: NoteValueBase;
  dots?: IntegerUnsigned;
};
export type NoteValueBase =
  | 'duplexMaxima'
  | 'maxima'
  | 'longa'
  | 'breve'
  | 'whole'
  | 'half'
  | 'quarter'
  | 'eighth'
  | '16th'
  | '32nd'
  | '64th'
  | '128th'
  | '256th'
  | '512th'
  | '1024th'
  | '2048th'
  | '4096th';
export type Time = GlobalAttrs & {
  count: PositiveInteger;
  display?: TimeSignatureDisplay;
  unit: TimeSignatureUnit;
};
export type PositiveInteger = number;
export type TimeSignatureDisplay = 'common' | 'cut';
export type TimeSignatureUnit = 1 | 2 | 4 | 8 | 16 | 32 | 64 | 128;
/**
 * This interface was referenced by `SoundsGlobal`'s JSON-Schema definition
 * via the `patternProperty` "^.*$".
 */
export type Sound = GlobalAttrs & {
  midiNumber?: MidiNumber;
  name?: String;
};
export type MidiNumber = number;
export type SystemLayout = GlobalAttrs & {
  content: SystemLayoutContent;
};
// ...
export type StaffNumber = number;
export type StemDirection = 'up' | 'down';
export type VoiceName = string;
export type SystemLayoutContent = (StaffGroup | Staff)[];
export type Mnx = GlobalAttrs & {
  support?: Support;
  version: VersionNumber;
};
export type Support = GlobalAttrs & {
  useAccidentalDisplay?: boolean;
  useBeams?: boolean;
};
export type VersionNumber = number;
export type Part = GlobalAttrs & {
  kit?: Kit;
  measures: PartMeasure[];
  name?: PartName;
  shortName?: PartShortName;
  smuflFont?: SmuflFont;
  staves?: StaffCount;
  transposition?: PartTransposition;
};
/**
 * This interface was referenced by `Kit`'s JSON-Schema definition
 * via the `patternProperty` "^.*$".
 */
export type KitComponent = GlobalAttrs & {
  name?: String;
  sound?: Id;
  staffPosition: StaffPosition;
};
export type StaffPosition = number;
export type PartMeasure = GlobalAttrs & {
  arpeggios?: Arpeggio[];
  beams?: BeamList;
  clefs?: PositionedClef[];
  dynamics?: DynamicGroup[];
  measureRepeat?: MeasureRepeat;
  nonArpeggios?: NonArpeggio[];
  ottavas?: Ottava[];
  sequences: Sequence[];
  staffConfigs?: PositionedStaffConfig[];
};
export type Arpeggio = GlobalAttrs & {
  arrow?: boolean;
  direction?: UpDownAuto;
  position: RhythmicPosition;
  span: IdPair;
};
export type Beam = GlobalAttrs & {
  beams?: BeamList;
  direction?: BeamHookDirection;
  events: Id[];
};
export type BeamHookDirection = 'left' | 'right' | 'auto';
export type BeamList = Beam[];
export type PositionedClef = GlobalAttrs & {
  clef: Clef;
  position?: RhythmicPosition;
  staff?: StaffNumber;
};
export type Clef = GlobalAttrs & {
  color?: SimpleColor;
  glyph?: SmuflGlyph;
  hide?: boolean;
  octave?: OttavaAmountOrZero;
  showOctave?: boolean;
  sign: ClefSign;
  staffPosition: StaffPosition;
};
export type SimpleColor = string;
export type OttavaAmountOrZero = 1 | 2 | -1 | -2 | 3 | -3 | 0;
export type ClefSign = 'C' | 'F' | 'G' | 'P';
export type DynamicGroup = GlobalAttrs & {
  accentPrefix?: DynamicPrefix;
  accentSuffix?: DynamicSuffix;
  end?: MeasureRhythmicPosition;
  glyphs?: SmuflGlyph[];
  placement?: MultiStaffPlacement;
  position: RhythmicPosition;
  prefix?: String;
  relativeValue?: RelativeDynamicValue;
  residualValue?: DynamicValue;
  staff?: StaffNumber;
  staffEnd?: StaffNumber;
  suffix?: String;
  type: DynamicGroupType;
  value?: DynamicValue;
  visuallyContinues?: Id;
  voice?: VoiceName;
  wedgeType?: WedgeType;
};
export type DynamicPrefix = 's' | 'r' | '';
export type DynamicSuffix = 'z' | '';
export type MeasureRhythmicPosition = GlobalAttrs & {
  measure: Id;
  position: RhythmicPosition;
};
export type MultiStaffPlacement = 'above' | 'auto' | 'below' | 'between';
export type RelativeDynamicValue = 'louder' | 'softer';
export type DynamicValue =
  | 'ppp'
  | 'pp'
  | 'p'
  | 'mp'
  | 'mf'
  | 'f'
  | 'ff'
  | 'fff'
  | 'n'
  | 'pppp'
  | 'ppppp'
  | 'ffff'
  | 'fffff'
  | 'pppppp'
  | 'ffffff';
export type DynamicGroupType = 'immediate' | 'gradual' | 'relative' | 'accent';
export type WedgeType = 'increasing' | 'decreasing';
export type MeasureRepeat = GlobalAttrs & {
  counter?: MeasureRepeatCounter;
  displayNumber?: YesNoAuto;
  number: MeasureRepeatCount;
  staffPosition?: StaffPosition;
};
export type MeasureRepeatCounter = GlobalAttrs & {
  count: PositiveInteger;
  placement?: MultiStaffPlacement;
};
export type YesNoAuto = 'yes' | 'no' | 'auto';
export type MeasureRepeatCount = number;
export type NonArpeggio = GlobalAttrs & {
  position: RhythmicPosition;
  span: IdPair;
};
export type Ottava = GlobalAttrs & {
  end: MeasureRhythmicPosition;
  placement?: Placement;
  position: RhythmicPosition;
  staff?: StaffNumber;
  value: OttavaAmount;
  voice?: VoiceName;
};
export type OttavaAmount = 1 | 2 | -1 | -2 | 3 | -3;
export type Sequence = GlobalAttrs & {
  content: SequenceContent;
  directionHint?: DirectionHint;
  fullMeasure?: FullMeasureRest;
  staff?: StaffNumber;
  voice?: VoiceName;
};
export type Event = GlobalAttrs & {
  duration: NoteValue;
  fermata?: Fermata;
  kitNotes?: KitNote[];
  lyrics?: Lyrics;
  markings?: EventMarkings;
  notes?: Note[];
  rest?: Rest;
  slurs?: Slur[];
  staff?: StaffNumber;
  stemDirection?: StemDirection;
  type?: 'event';
};
export type KitNote = GlobalAttrs & {
  kitComponent: Id;
  perform?: PerformOptions;
  staff?: StaffNumber;
  ties?: TieList;
};
export type PerformOptions = GlobalAttrs & {};
export type Tie = GlobalAttrs & {
  lv?: boolean;
  side?: SlurSide;
  target?: Id;
  targetType?: TieTargetType;
};
export type SlurSide = 'up' | 'down' | 'auto';
export type TieTargetType = 'nextNote' | 'crossVoice' | 'arpeggio' | 'crossJump';
export type TieList = Tie[];
export type Lyrics = GlobalAttrs & {
  lines?: EventLyricLines;
};
/**
 * This interface was referenced by `EventLyricLines`'s JSON-Schema definition
 * via the `patternProperty` "^.*$".
 */
export type EventLyricLine = GlobalAttrs & {
  text: String;
  type?: EventLyricLineType;
};
export type EventLyricLineType = 'start' | 'middle' | 'end' | 'whole';
export type EventMarkings = GlobalAttrs & {
  accent?: Accent;
  bowDirection?: BowDirection;
  breath?: BreathMark;
  caesura?: Caesura;
  softAccent?: SoftAccent;
  spiccato?: Spiccato;
  staccatissimo?: Staccatissimo;
  staccato?: Staccato;
  stress?: StressMarking;
  strongAccent?: StrongAccent;
  tenuto?: Tenuto;
  tremolo?: TremoloSingle;
  unstress?: UnstressMarking;
};
export type Accent = GlobalAttrs & {
  placement?: Placement;
};
export type BowDirection = GlobalAttrs & {
  direction: UpDown;
  placement?: Placement;
};
export type UpDown = 'up' | 'down';
export type BreathMark = GlobalAttrs & {
  placement?: Placement;
  symbol?: BreathMarkSymbol;
};
export type BreathMarkSymbol = 'comma' | 'tick' | 'upbow' | 'salzedo' | 'auto';
export type Caesura = GlobalAttrs & {
  marks?: number;
  shape?: CaesuraShape;
};
export type CaesuraShape = 'normal' | 'thick' | 'short' | 'curved';
export type SoftAccent = GlobalAttrs & {
  placement?: Placement;
};
export type Spiccato = GlobalAttrs & {
  placement?: Placement;
};
export type Staccatissimo = GlobalAttrs & {
  placement?: Placement;
};
export type Staccato = GlobalAttrs & {
  placement?: Placement;
};
export type StressMarking = GlobalAttrs & {
  placement?: Placement;
};
export type StrongAccent = GlobalAttrs & {
  placement?: Placement;
  pointing?: UpDownAuto;
};
export type Tenuto = GlobalAttrs & {
  placement?: Placement;
};
export type TremoloSingle = GlobalAttrs & {
  marks: number;
  placement?: Placement;
};
export type UnstressMarking = GlobalAttrs & {
  placement?: Placement;
};
export type Note = GlobalAttrs & {
  accidentalDisplay?: AccidentalDisplay;
  perform?: PerformOptions;
  pitch: Pitch;
  staff?: StaffNumber;
  ties?: TieList;
  written?: Written;
};
export type AccidentalDisplay = GlobalAttrs & {
  enclosure?: AccidentalEnclosure;
  force?: boolean;
  show: boolean;
};
export type AccidentalEnclosure = GlobalAttrs & {
  symbol: AccidentalEnclosureSymbol;
};
export type AccidentalEnclosureSymbol = 'parentheses' | 'brackets';
export type Pitch = GlobalAttrs & {
  alter?: Alter;
  octave: Octave;
  step: Step;
};
export type Alter = number;
export type Octave = number;
export type Step = 'A' | 'B' | 'C' | 'D' | 'E' | 'F' | 'G';
export type Written = GlobalAttrs & {
  diatonicDelta?: IntegerSigned;
};
export type IntegerSigned = number;
export type Rest = GlobalAttrs & {
  staffPosition?: StaffPosition;
};
export type Slur = GlobalAttrs & {
  endNote?: Id;
  lineType?: LineType;
  side?: SlurSide;
  sideEnd?: SlurSide;
  startNote?: Id;
  target: Id;
};
export type LineType = 'dashed' | 'dotted' | 'solid' | 'wavy';
export type Grace = GlobalAttrs & {
  color?: Color;
  content: Event[];
  graceType?: GraceType;
  slash?: boolean;
  type: 'grace';
};
export type GraceType = 'makeTime' | 'stealFollowing' | 'stealPrevious';
export type Tuplet = GlobalAttrs & {
  bracket?: YesNoAuto;
  content: SequenceContent;
  inner: NoteValueQuantity;
  outer: NoteValueQuantity;
  placement?: Placement;
  showNumber?: TupletDisplaySetting;
  showValue?: TupletDisplaySetting;
  staff?: StaffNumber;
  type: 'tuplet';
};
export type NoteValueQuantity = GlobalAttrs & {
  duration: NoteValue;
  multiple: PositiveInteger;
};
export type TupletDisplaySetting = 'noNumber' | 'inner' | 'both';
export type Space = GlobalAttrs & {
  duration: Fraction;
  type: 'space';
};
export type MultiNoteTremolo = GlobalAttrs & {
  content: Event[];
  individualDuration?: NoteValue;
  marks: number;
  outer: NoteValueQuantity;
  type: 'tremolo';
};
export type SequenceContent = (Event | Grace | Tuplet | Space | MultiNoteTremolo)[];
export type DirectionHint = 'upper' | 'lower' | 'auto';
export type FullMeasureRest = GlobalAttrs & {
  fermata?: Fermata;
  staffPosition?: StaffPosition;
  visualDuration?: NoteValue;
};
export type PositionedStaffConfig = GlobalAttrs & {
  config: StaffConfig;
  position?: RhythmicPosition;
  staff?: StaffNumber;
};
export type StaffConfig = GlobalAttrs & {
  lines?: IntegerUnsigned;
};
export type PartName = string;
export type PartShortName = string;
export type SmuflFont = string;
export type StaffCount = number;
export type PartTransposition = GlobalAttrs & {
  interval: Interval;
  keyFifthsFlipAt?: IntegerSigned;
  prefersWrittenPitches?: boolean;
};
export type Interval = GlobalAttrs & {
  halfSteps: IntegerSigned;
  staffDistance: IntegerSigned;
};
export type Score = GlobalAttrs & {
  layout?: Id;
  multimeasureRests?: MultimeasureRest[];
  name: ScoreName;
  pages?: Page[];
  useWritten?: boolean;
};
export type MultimeasureRest = GlobalAttrs & {
  duration: MeasureCount;
  label?: String;
  start: Id;
};
export type MeasureCount = number;
export type ScoreName = string;
export type Page = GlobalAttrs & {
  layout?: Id;
  systems: System[];
};
export type System = GlobalAttrs & {
  layout?: Id;
  layoutChanges?: LayoutChange[];
  measure: Id;
};
export type LayoutChange = GlobalAttrs & {
  layout: Id;
  location: MeasureRhythmicPosition;
};

export interface GlobalAttrs {
  _c?: String;
  _x?: VendorExtensions;
  id?: Id;
}
export interface VendorExtensions {
  [k: string]: VendorDict;
}
/**
 * This interface was referenced by `VendorExtensions`'s JSON-Schema definition
 * via the `patternProperty` "^.*$".
 */
export interface VendorDict {}
export interface LyricLinesMetadata {
  [k: string]: LyricLineMetadata;
}
export interface SoundsGlobal {
  [k: string]: Sound;
}
export interface Kit {
  [k: string]: KitComponent;
}
export interface IdPair {
  end: Id;
  start: Id;
}
export interface EventLyricLines {
  [k: string]: EventLyricLine;
}